% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/healthcarerisk.R
\name{charlsonindex}
\alias{charlsonindex}
\title{Calculating the Charlson Index}
\usage{
charlsonindex(abstract_data)
}
\arguments{
\item{abstract_data}{a data.frame requiring 25 ICD-10 diagnosis codes per row with naming convention diag_code_01 - diag_code_25
followed by 25 diagnosis types (as per CIHI coding standards) with naming convetion diag_type_01 - diag_type_25. If less than 25 diagnoses were codes then excess columns should be loaded with NA or an empty string}
}
\value{
A data.frame with all columns from abstract_data with with Charlson Index score appended, with values between 0 and 24 (0 is no comorbidities, 24 is all comorbidities)
}
\description{
This function applies the updated Charlson Index to discharge abstract data set. The Charlson Index is used to weight the severity of
comorbidities required for risk adjusting mortality ratio, readmissions, and other quality indicators.
Codes and updated weights were taken from two Hude Quan articles referenced below.
}
\details{
There are 12 comorbidities with contribute to Charlson Index with the following weights:

Congestive heart failure - 2; Dementia - 2; Chronic pulmonary disease - 1; Rheamatologic disease - 1;
Mild liver disease - 2; Diabetes with chronic complications - 1; Hemiplagia or paraplagia - 2;
Renal disease - 1; Any malignancy, including leukemia and lymphoma - 2; Moderate or severe liver disease - 4; Metastatic solid tumour - 6;
AIDS/HIV - 4

The maximum weight is 24. The following combinations are mutually exclusive: Mild liver disease & moderate/server liver disease; any malignancy & metastatic solid tumour

Only diagnosis types 1, W, Y, X are used in calculating Charlson Index.
}
\author{
Josh Morel
}
\references{
Hude Quan. Coding algorithms for defining comorbidities in ICD-9-CM and ICD-10 administrative data. Med Care, 43(11):1130-9.

Hude Quan. Updated and Validating the Charlson Comorbidity Index and Score for Risk Adjustment in Hospital Discharge Using Data from 6 Countries. American Journal of Epidemiology. 2011.

Canadian Institute of Health Information (CIHI). DAD Abstracting Manual, 2015-2016 Edition.
}

